// src/data/FilterData.d.ts
export declare const FilterData: {
    title: string;
    classname: string;
    sortOptions: {
        name: string;
        href: string;
        current: boolean;
    }[];
    filters: {
        id: string;
        name: string;
        options: {
            value: string;
            label: string;
            checked: boolean;
        }[];
    }[];
};
//# sourceMappingURL=FilterData.d.ts.map